+++ corewar/src/instructions/cw_sti.c
@@ -6,7 +6,7 @@
 /*   By: tdeborde <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/26 15:08:22 by tdeborde          #+#    #+#             */
-/*   Updated: 2018/04/20 12:37:30 by blefeuvr         ###   ########.fr       */
+/*   Updated: 2018/04/20 14:48:04 by tdeborde         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ int		cw_sti_param(t_vm *vm, t_process *process, int param[3], int ret[3])
 			param[i] = cw_get_4(process->registries[param[i] - 1]);
 		else if (ret[i] == 2)
 			param[i] = cw_get_2(&(vm->vm[(process->pc + process->entrypoint
-							- offset + ((short)param[i] % IDX_MOD)) % MEM_SIZE]));
+							- offset + (param[i] % IDX_MOD)) % MEM_SIZE]));
 		ret[i] = ret[i] == 4 ? 2 : ret[i];
 		ocp = ocp << 2;
 	}
